+++ torchao/quantization/quantize_/workflows/__init__.py
@@ -1,12 +1,12 @@
 from .float8.float8_packing_format import (
     Float8PackingFormat,
 )
-from .float8.float8_semi_sparse_tensor import (
-    Sparse2x4CUTLASSFloat8Tensor,
-)
 from .float8.float8_tensor import (
     Float8Tensor,
     QuantizeTensorToFloat8Kwargs,
+)
+from .float8.sparse_2x4_cutlass_float8_tensor import (
+    Sparse2x4CUTLASSFloat8Tensor,
 )
 from .int4.int4_choose_qparams_algorithm import Int4ChooseQParamsAlgorithm
 from .int4.int4_marlin_sparse_tensor import (
